+++ system_diagram.png
